Much cleaner implementation.
We now have the user pass a QuasiNewtonModel object to AugLag, which is
then converted to a SlackModel internally. The SlackModel class already
implements the correct Hessian-vector product for the matrix structure.
This avoids the definition of a whole bunch of specialized classes. Only
TRON is subclassed to compute the correct set of update vectors. The full
quasi-Newton approximation of the AugmentedLagrangian model is
still available upon request.